--- tests/test_example.py
@@ -1,5 +1,0 @@
-import pytest
-
-
-def test_example():
-    assert True
